refactor: adjust command domains
