+++ src/commands.ts
@@ -4,32 +4,32 @@ import * as vscode from 'vscode';
 import { registerFileSelectionCommands } from '@commands/files/fileSelectionCommands';
 import { registerLatexdiffCommands } from '@commands/latex/latexdiffCommands';
 import { registerGitCommands } from '@commands/git/gitCommands';
-import { registerPackCommands } from './commands/packCommands';
-import { registerCleanCommands } from './commands/cleanCommands';
-import { registerMergeCommands } from './commands/mergeCommands';
-import { registerExecuteCommand } from './commands/executeCommand';
+import { registerPackCommands } from '@commands/housekeeping/packCommands';
+import { registerCleanCommands } from '@commands/housekeeping/cleanCommands';
+import { registerMergeCommands } from '@commands/agent/mergeCommands';
+import { registerExecuteCommand } from '@commands/agent/executeCommand';
 import { TeXRAViewProvider } from './ViewProvider';
 import { registerLatexCommands } from '@commands/latex/latexCommands';
-import { registerImageCommands } from './commands/imageCommands';
-import { registerFigureCommands } from './commands/figCommands';
-import { registerTestCommands } from './commands/testCommands';
-import { registerXmlCommands } from './commands/xmlCommands';
-import { registerYamlCommands } from './commands/yamlCommands';
-import { registerAgentCommands } from './commands/agentCommands';
-import { registerAgentCreatorCommands } from '@commands/agentCreatorCommands';
+import { registerImageCommands } from '@commands/latex/imageCommands';
+import { registerFigureCommands } from '@commands/latex/figCommands';
+import { registerTestCommands } from '@commands/system/testCommands';
+import { registerXmlCommands } from '@commands/system/xmlCommands';
+import { registerYamlCommands } from '@commands/system/yamlCommands';
+import { registerAgentCommands } from '@commands/agent/agentCommands';
+import { registerAgentCreatorCommands } from '@commands/agent/agentCreatorCommands';
 import { registerApiKeyCommands } from '@commands/api/apiKeyCommands';
-import { registerStateRestoreCommand } from './commands/stateRestoreCommand';
-import { registerTextEditorCommands } from './commands/textEditorCommands';
+import { registerStateRestoreCommand } from '@commands/history/stateRestoreCommand';
+import { registerTextEditorCommands } from '@commands/system/textEditorCommands';
 import { registerLinterCommands } from '@commands/latex/linterCommands';
 import { registerWolframAlphaCommands } from '@commands/wolfram/wolframAlphaCommands';
 import { registerWolframScriptCommands } from '@commands/wolfram/wolframScriptCommands';
 import { registerHistoryCommands } from '@commands/history/historyCommands';
-import { registerArXivCommands } from './commands/arXivCommands';
-import { registerCompareCommands } from './commands/compareCommands';
-import { registerHelpCommands } from './commands/helpCommands';
+import { registerArXivCommands } from '@commands/latex/arXivCommands';
+import { registerCompareCommands } from '@commands/latex/compareCommands';
+import { registerHelpCommands } from '@commands/system/helpCommands';
 import { registerProgressViewCommands } from '@commands/progress/progressViewCommands';
 import { registerOpenFileCommands } from '@commands/files/openFileCommands';
-import { registerSettingsCommands } from './commands/settingsCommands';
+import { registerSettingsCommands } from '@commands/system/settingsCommands';
 
 import * as logger from '@logger/logUtils';
 
@@ -93,25 +93,25 @@ export function registerCommands(context: vscode.ExtensionContext) {
 export { fileSelectionCommands } from '@commands/files/fileSelectionCommands';
 export { latexdiffCommands } from '@commands/latex/latexdiffCommands';
 export { gitCommands } from '@commands/git/gitCommands';
-export { packCommands } from './commands/packCommands';
-export { mergeCommands } from './commands/mergeCommands';
-export { executeCommand } from './commands/executeCommand';
+export { packCommands } from '@commands/housekeeping/packCommands';
+export { mergeCommands } from '@commands/agent/mergeCommands';
+export { executeCommand } from '@commands/agent/executeCommand';
 export { latexCommands } from '@commands/latex/latexCommands';
-export { imageCommands } from './commands/imageCommands';
-export { figureCommands } from './commands/figCommands';
-export { testCommands } from './commands/testCommands';
-export { xmlCommands } from './commands/xmlCommands';
-export { yamlCommands } from './commands/yamlCommands';
+export { imageCommands } from '@commands/latex/imageCommands';
+export { figureCommands } from '@commands/latex/figCommands';
+export { testCommands } from '@commands/system/testCommands';
+export { xmlCommands } from '@commands/system/xmlCommands';
+export { yamlCommands } from '@commands/system/yamlCommands';
 export { apiKeyCommands } from '@commands/api/apiKeyCommands';
-export { stateRestoreCommand } from './commands/stateRestoreCommand';
-export { textEditorCommands } from './commands/textEditorCommands';
+export { stateRestoreCommand } from '@commands/history/stateRestoreCommand';
+export { textEditorCommands } from '@commands/system/textEditorCommands';
 export { linterCommands } from '@commands/latex/linterCommands';
 export { wolframAlphaCommands } from '@commands/wolfram/wolframAlphaCommands';
 export { wolframScriptCommands } from '@commands/wolfram/wolframScriptCommands';
 export { historyCommands } from '@commands/history/historyCommands';
-export { arXivCommands } from './commands/arXivCommands';
-export { compareCommands } from './commands/compareCommands';
-export { helpCommands } from './commands/helpCommands';
-export { agentCreatorCommands } from '@commands/agentCreatorCommands';
+export { arXivCommands } from '@commands/latex/arXivCommands';
+export { compareCommands } from '@commands/latex/compareCommands';
+export { helpCommands } from '@commands/system/helpCommands';
+export { agentCreatorCommands } from '@commands/agent/agentCreatorCommands';
 export { registerOpenFileCommands as openFileCommands } from '@commands/files/openFileCommands';
-export { settingsCommands } from './commands/settingsCommands';
+export { settingsCommands } from '@commands/system/settingsCommands';
