+++ src/commands/agent/mergeCommands.ts
