+++ src/commands/latex/compareCommands.ts
